Change how KNemo does polling
Support sub-second poll intervals, but only for specific values.
All updates triggered by the poll timer.

svn path=/trunk/extragear/network/knemo/; revision=1040924

+++ src/knemod/global.h
@@ -34,7 +34,7 @@
 struct GeneralData
 {
     int toolTipContent;
-    int pollInterval;
+    double pollInterval;
     int saveInterval;
     KUrl statisticsDir;
 };
